Remove TraceShow on interface array length.
Left behind this when I was debugging the interfaces function
(issue, as previously mentioned, was the lack of base case in the
recursive function).

An aside, enforcing functions to be Total would be great.
Alternatively, good reason to avoid using recursion for looping and
rely on constructs like fold instead.