Added Fields.
Running into the haskell weirdness of not allowing repeated structure
names. So access_field had to be modified with a prefix. Newer versions
of ghc have a way to change that behavior.

My understanding is the reason that happens is because all strcuture
components become named accessor functions. Why they didn't simply
namespace them to avoid this issue is beyond me.

So for example, running in ghci:
  > :t c_access_flags
  c_access_flags :: ClassFile -> Word16

I believe they could have simply done
  > :t ClassFile.c_access_flags
  ClassFile.c_access_flags :: ClassFile -> Word16

And everyone would have gone home happy. :|

So method refs is the only thing left!