Using Pretty Printing via pretty-simple
Simple, really nice addition. Main thing we need to do is take the raw
byte objects (in the attribute objects) and change the Show method on
them, as they currently dump the raw values to the terminal. :( This
should go back to the old 'hex escaped' version.

Also, we should really figure out when we need to decompile.