Switching to pretty-show for now.
So pretty-print seemed to be directly printing the raw bytestrings
rather than escaping them appropriately. Possibly a bug in their
library, but I was unable to figure out how their code worked exactly.
More investigation should be done.

One of the things I've realized is that attributes have additional
subformats, that is, their info[] arrays have structure to them that is
defined in the oracle docs.

To get feature completeness, we should decode these correctly.